feat: add code variables support
