+++ rust/cymbal/src/fingerprinting/mod.rs
@@ -139,6 +139,7 @@ mod test {
                 resolve_failure: None,
                 lang: "javascript".to_string(),
                 junk_drawer: None,
+                code_variables: None,
                 context: None,
                 release: None,
                 synthetic: false,
@@ -157,6 +158,7 @@ mod test {
                 resolve_failure: None,
                 lang: "javascript".to_string(),
                 junk_drawer: None,
+                code_variables: None,
                 context: None,
                 release: None,
                 synthetic: false,
@@ -177,6 +179,7 @@ mod test {
             resolve_failure: None,
             lang: "javascript".to_string(),
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             synthetic: false,
@@ -227,6 +230,7 @@ mod test {
                 resolve_failure: None,
                 lang: "javascript".to_string(),
                 junk_drawer: None,
+                code_variables: None,
                 context: None,
                 release: None,
                 synthetic: false,
@@ -245,6 +249,7 @@ mod test {
                 resolve_failure: None,
                 lang: "javascript".to_string(),
                 junk_drawer: None,
+                code_variables: None,
                 context: None,
                 release: None,
                 synthetic: false,
@@ -263,6 +268,7 @@ mod test {
                 resolve_failure: None,
                 lang: "javascript".to_string(),
                 junk_drawer: None,
+                code_variables: None,
                 context: None,
                 release: None,
                 synthetic: false,
@@ -307,6 +313,7 @@ mod test {
             resolve_failure: None,
             lang: "javascript".to_string(),
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             synthetic: false,
@@ -326,6 +333,7 @@ mod test {
             resolve_failure: None,
             lang: "javascript".to_string(),
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             synthetic: false,
