+++ rust/cymbal/src/langs/go.rs
@@ -42,6 +42,7 @@ impl From<&RawGoFrame> for Frame {
             release: None,
             suspicious: false,
             module: None,
+            code_variables: None,
         }
     }
 }
