+++ rust/cymbal/src/langs/hermes.rs
@@ -125,6 +125,7 @@ impl From<(&RawHermesFrame, HermesError)> for Frame {
             resolve_failure: Some(err.to_string()),
             synthetic: frame.meta.synthetic,
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             suspicious: false,
@@ -158,6 +159,7 @@ impl From<(&RawHermesFrame, Token<'_>, Option<String>)> for Frame {
             resolve_failure: None,
             synthetic: frame.meta.synthetic,
             junk_drawer: None,
+            code_variables: None,
             context: get_token_context(&token, token.get_src_line() as usize),
             release: None,
             suspicious: false,
