+++ rust/cymbal/src/langs/java.rs
@@ -142,6 +142,7 @@ impl<'a> From<(&'a RawJavaFrame, StackFrame<'a>)> for Frame {
             resolved: true,
             resolve_failure: None,
             junk_drawer: None,
+            code_variables: None,
             release: None,
             synthetic: raw.meta.synthetic,
             context: None,
@@ -169,6 +170,7 @@ impl From<(&RawJavaFrame, ProguardError)> for Frame {
             resolved: false,
             resolve_failure: Some(error.to_string()),
             junk_drawer: None,
+            code_variables: None,
             release: None,
             synthetic: raw.meta.synthetic,
             context: None,
