+++ rust/cymbal/src/langs/js.rs
@@ -210,6 +210,7 @@ impl From<(&RawJSFrame, SourceLocation<'_>)> for Frame {
             resolved: true,
             resolve_failure: None,
             junk_drawer: None,
+            code_variables: None,
             context: get_sourcelocation_context(&token),
             release: None,
             synthetic: raw_frame.meta.synthetic,
@@ -258,6 +259,7 @@ impl From<(&RawJSFrame, JsResolveErr, &FrameLocation)> for Frame {
             // why we thought a frame wasn't minified, they can see the error message
             resolve_failure: Some(err.to_string()),
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             synthetic: raw_frame.meta.synthetic,
@@ -298,6 +300,7 @@ impl From<&RawJSFrame> for Frame {
             resolved: true, // Without location information, we're assuming this is not minified
             resolve_failure: None,
             junk_drawer: None,
+            code_variables: None,
             context: None,
             release: None,
             synthetic: raw_frame.meta.synthetic,
