+++ rust/cymbal/src/langs/node.rs
@@ -166,6 +166,7 @@ impl From<&RawNodeFrame> for Frame {
             synthetic: raw.meta.synthetic,
             suspicious: false,
             module: raw.module.clone(),
+            code_variables: None,
         }
     }
 }
@@ -204,6 +205,7 @@ impl From<(&RawNodeFrame, SourceLocation<'_>)> for Frame {
             resolved: true,
             resolve_failure: None,
             junk_drawer: None,
+            code_variables: None,
             context: get_sourcelocation_context(&location),
             release: None,
             synthetic: raw_frame.meta.synthetic,
@@ -248,6 +250,7 @@ impl From<(&RawNodeFrame, JsResolveErr)> for Frame {
             // why we thought a frame wasn't minified, they can see the error message
             resolve_failure: Some(resolve_err.to_string()),
             junk_drawer: None,
+            code_variables: None,
             context: raw_frame.get_context(),
             release: None,
             synthetic: raw_frame.meta.synthetic,
